modernize code
Conflicts:

	fe-aqua/ChatWindow.mm
	fe-aqua/SGTabView.mm

+++ fe-aqua/MenuMaker.mm
@@ -458,7 +458,7 @@ static MenuMaker *defaultMenuMaker;
 
 - (NSString *)stripImageFromTitle:(NSString *)title icon:(NSString **)icon
 {
-    int length;
+    NSInteger length;
     // stringByReplacingOccurrencesOfString is not available on 10.4.
     //    title = [title stringByReplacingOccurrencesOfString:@"_" withString:@""];
     NSMutableString *mTitle = [title mutableCopy];
